+++ crates/dispatch2/CHANGELOG.md
@@ -45,6 +45,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
   - `dispatch_wait`.
   - `dispatch_write`.
 - Added `#[must_use]` in bindings where the source uses it.
+- Added `Queue::main` for fetching the main queue.
+- Moved `run_on_main` and `MainThreadBound` from `objc2-foundation` to this crate.
 
 ### Changed
 - Moved to the `objc2` project.